pyproject.toml
[project]
name = "ai-interview-pipeline"
version = "0.1.0"
description = "AI算法工程师面经知识库处理流水线"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "rapidfuzz>=3.0.0",
    "pandas>=2.0.0",
    "duckdb>=0.10.0",
    "orjson>=3.9.0",
    "tenacity>=8.0.0",
    "typer>=0.9.0",
    "python-dotenv>=1.0.0",
    "openpyxl>=3.1.0",
    "sentence-transformers>=2.2.0",
    "FlagEmbedding>=1.2.0",
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
]

[project.scripts]
interview-pipeline = "src.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
